+++ src/theme.js
@@ -1,5 +1,4 @@
 import { DefaultTheme } from 'react-native-paper';
-
 export const theme = {
   ...DefaultTheme,
   roundness: 27,
